Add Unit Test for eCR liquid templaste (#19)
This PR adds unit tests for the ContactPoint data type in the eCR templates. It is used as a proof of concept for unit testing liquid templates in general.

A base class is used to abstract out the guts of the testing. A user just needs to inherit from that class and then use the ConvertCheckLiquidTemplate function with the path to the template they are testing, the attributes that should be passed to the template, and the expected string post-rendering (sans newlines and tabs for sanity reasons).